default.nix: Benchmark 3 different LuaJIT versions

+++ default.nix
@@ -1,39 +1,44 @@
 { pkgs ? import <nixpkgs> {},
-  luajitName ? "unknown",
-  luajitSrc,
-  testsuiteSrc ? null,
+  luajitAname ? "unknown",
+  luajitAsrc,
+  luajitBname,
+  luajitBsrc,
+  luajitCname,
+  luajitCsrc,
+  testsuiteSrc,
+  hardware ? null,
   benchmarkRuns ? 1 }:
 
 with pkgs;
 with stdenv;
 
-rec {
-  # Build LuaJIT for testing
-  luajit = mkDerivation {
+# LuaJIT build derivation
+let buildLuaJIT = luajitName: luajitSrc: mkDerivation {
     name = "luajit-${luajitName}";
+    version = luajitName;
     src = luajitSrc;
     enableParallelBuilding = true;
     installPhase = ''
       make install PREFIX=$out
-      ln -s $out/bin/luajit-* $out/bin/luajit
+      if [ ! -e $out/bin/luajit ]; then
+        ln -s $out/bin/luajit-* $out/bin/luajit
+      fi
     '';
-  };
+  }; in
 
-  # Build LuaJIT with -Werror to check for compile issues
-  luajit-compile-Werror = lib.overrideDerivation luajit (old: {
-    name = "luajit-Werror";
-    phases = "unpackPhase buildPhase";
-    NIX_CFLAGS_COMPILE = "-Werror";
-  });
-
-  # Run the standard LuaJIT benchmarks many times and produce a CSV file.
-  benchmarks = mkDerivation {
-    name = "luajit-benchmarks";
+# LuaJIT benchmark run derivatin
+# Run the standard LuaJIT benchmarks many times and produce a CSV file.
+let benchmarkLuaJIT = luajitName: luajitSrc:
+  let luajit = (buildLuaJIT luajitName luajitSrc); in
+  mkDerivation {
+    name = "luajit-${luajitName}-benchmarks";
     src = testsuiteSrc;
+    # Force consistent hardware
+    requiredHardwareFeatures = if hardware != null then [hardware] else [];
     buildInputs = [ luajit linuxPackages.perf ];
     buildPhase = ''
       PATH=luajit/bin:$perf/bin:$PATH
-      # Run multiple sets of benchmarks
+      # Run multiple iterations of the benchmarks
       for run in $(seq 1 ${toString benchmarkRuns}); do
         echo "Run $run"
         mkdir -p result/$run
@@ -53,12 +58,11 @@ rec {
       # Copy the raw perf output for reference
       cp -r result $out
       # Create a CSV file
-      echo "luajit,benchmark,run,instructions,cycles" > $out/bench.csv
       for resultdir in result/*; do
         run=$(basename $resultdir)
         # Create the rows based on the perf logs
         for result in $resultdir/*.perf; do
-          luajit=${luajit.name}
+          luajit=${luajit.version}
           benchmark=$(basename -s.perf -a $result)
           instructions=$(awk -F, -e '$3 == "instructions" { print $4; }' $result)
           cycles=$(      awk -F, -e '$3 == "cycles"       { print $4; }' $result)
@@ -66,7 +70,19 @@ rec {
         done
       done
     '';
-  };
+  }; in
+
+rec {
+  # Build LuaJIT with -Werror to check for compile issues
+  luajit-compile-Werror = lib.overrideDerivation luajit (old: {
+    name = "luajit-Werror";
+    phases = "unpackPhase buildPhase";
+    NIX_CFLAGS_COMPILE = "-Werror";
+  });
+
+  benchmarksA = (benchmarkLuaJIT luajitAname luajitAsrc);
+  benchmarksB = (benchmarkLuaJIT luajitBname luajitBsrc);
+  benchmarksC = (benchmarkLuaJIT luajitCname luajitCsrc);
 
   benchmarkResults = mkDerivation {
     name = "benchmark-results";
@@ -75,17 +91,18 @@ rec {
       source $stdenv/setup
       # Get the CSV file
       mkdir -p $out/nix-support
-      cp ${benchmarks}/bench.csv $out/
+      echo "luajit,benchmark,run,instructions,cycles" > bench.csv
+      cat ${benchmarksA}/bench.csv ${benchmarksB}/bench.csv ${benchmarksC}/bench.csv \
+          >> bench.csv
+      cp bench.csv $out
       echo "file CSV $out/bench.csv" >> $out/nix-support/hydra-build-products
       # Generate the report
       cp ${./benchmark-results.Rmd} benchmark-results.Rmd
-      cp ${benchmarks}/bench.csv .
       cat benchmark-results.Rmd
       echo "library(rmarkdown); render('benchmark-results.Rmd')"| R --no-save
       cp benchmark-results.html $out
       echo "file HTML $out/benchmark-results.html"  >> $out/nix-support/hydra-build-products
     '';
   };
-
 }
